+++ services/devstar/v1/model/ShowRepositoryByCloudIdeResponse.ts
@@ -1,0 +1,13 @@
+
+import { SdkResponse } from "@huaweicloud/huaweicloud-sdk-core/SdkResponse";
+
+export class ShowRepositoryByCloudIdeResponse extends SdkResponse {
+    public url?: string;
+    public constructor() { 
+        super();
+    }
+    public withUrl(url: string): ShowRepositoryByCloudIdeResponse {
+        this['url'] = url;
+        return this;
+    }
+}